// src/types/Compete.ts
export type Compete = {
  subject: string;
  industrySector: string;
  otherIndustrySector: string;
  numberOfEmployee: number;
  totalRevenue: number;
  employmentMonth: number;
  lastSalary: number;
  employmentEndDate: string;
  jobTitle: string;
  otherJobTitle: string;
  employmentType: string;
  nonCompetePeriod: number;
  compensation: string;
  salaryPercentage: number;
  compensationSource: string;
  reasonsForClause: [string];
  otherReasonsForClause: string;
  prohibitedActivities: [string];
  otherProhibitedActivities: string;
  service: string;
  geoGraphical:[string];
  industrySectorCompany: string;
  explicitly: string;
  city: string;
  departmental: string;
  otherGeographical: string;
  employeeLocation: number;
  companyLocation: number;
  penaltyClause: string;
  penaltyAmount: number;
  judicialRequest: string;
  lastMonthlySalary: number;
  nonCompeteDuration: number;
  potentialLoss: number;
  breach: string;
  clauseBelieved: string;
  otherClauseBelieved: string;
  newCompany: string;
  position: string;
  companyCompete: string;
  newBusiness: string;
  typeOfBusiness: string;
  businessCompete: string;
  startJob: string;
  violationEvidence: [string];
  otherViolation: string;
  nonCompetitionClause: [string];
  otherClause: string;
  allegedViolation: [string];
  otherAlleged: string;
  agreement: string;
  negotiate: string;
  whenNegotiate: [string];
  otherDetails: string;

  progress: number;
  nextProgress: number;
  checked: boolean;

  user: string;
  caseId: string;
};